+++ cashu-wallet-protocol/src/main/java/xyz/tcheeric/cashu/wallet/proto/service/impl/DefaultDLEQVerificationService.java
@@ -134,7 +134,7 @@ public final class DefaultDLEQVerificationService implements DLEQVerificationSer
                 dleq.getE(),
                 dleq.getS(),
                 dleq.getR(),
-                proof.getSecret().toBytes(),
+                proof.getSecret().toString(),
                 unblindedSignaturePoint,
                 mintPublicKeyPoint
             );